dcrpg: add index inventory, missing/exists functions, and tests
main: force reindex and addresses spend update when missing indexes

Also make (*ChainDB).Height return an int64 instead of uint64 since -1
is the special value indicating empty tables.  Note that HeightDB()
still returns uint64 since it's error indicates empty (sql.ErrNoRows).
In (*ChainDB).SyncChainDB force reindex if lastBlock == -1.

remove old ChainDB.GetHeight

New ChainDB.GetHeight wraps GetHeight for middleware DataSource.
Change HeightDB to return an int64 instead of uint64.  A value of -1 is
returned when err == sql.ErrNoRows.
Instead of ChainDB.GetHeight, use HeightDB or Height.
Remove some ununsed queries.
docs.

+++ api/apiroutes.go
@@ -103,10 +103,10 @@ type DataSourceAux interface {
 	TxHistoryData(address string, addrChart dbtypes.HistoryChart,
 		chartGroupings dbtypes.TimeBasedGrouping) (*dbtypes.ChartsData, error)
 	TicketPoolVisualization(interval dbtypes.TimeBasedGrouping) (
-		*dbtypes.PoolTicketsData, *dbtypes.PoolTicketsData, *dbtypes.PoolTicketsData, uint64, error)
+		*dbtypes.PoolTicketsData, *dbtypes.PoolTicketsData, *dbtypes.PoolTicketsData, int64, error)
 	AgendaVotes(agendaID string, chartType int) (*dbtypes.AgendaVoteChoices, error)
 	AddressTxIoCsv(address string) ([][]string, error)
-	Height() uint64
+	Height() int64
 	AllAgendas() (map[string]dbtypes.MileStone, error)
 }
 
@@ -923,7 +923,7 @@ func (c *appContext) getTicketPoolCharts(w http.ResponseWriter, r *http.Request)
 	mp := c.BlockData.GetMempoolPriceCountTime()
 
 	response := &apitypes.TicketPoolChartsData{
-		ChartHeight: height,
+		ChartHeight: uint64(height),
 		TimeChart:   timeChart,
 		PriceChart:  priceChart,
 		DonutChart:  donutChart,
@@ -964,7 +964,7 @@ func (c *appContext) getTicketPoolByDate(w http.ResponseWriter, r *http.Request)
 	}
 
 	tpResponse := struct {
-		Height    uint64                   `json:"height"`
+		Height    int64                    `json:"height"`
 		TimeChart *dbtypes.PoolTicketsData `json:"time_chart"`
 	}{
 		height,
